Moved Image to logged_data contracts
